+++ docs/user/release_notes.rst
@@ -75,6 +75,7 @@ Dependency changes
   (:issue:`2700`).
 * pillow is no longer required (:issue:`2700`).
 * six is no longer required for the client or server (:issue:`2759`).
+* kitchen is no longer required.
 
 
 Server upgrade instructions